requirements.txt
python-telegram-bot
transformers
torch
accelerate
python-dotenv